+++ minfraud/models.py
@@ -708,12 +708,28 @@ class Subscores(object):
 
       :type: float | None
 
+      .. deprecated:: 1.8.0
+        Deprecated effective August 29, 2019. This subscore will default to 1
+        and will be removed in a future release. The user tenure on email is
+        reflected in the email address subscore output.
+
+      .. seealso::
+        :py:attr:`minfraud.models.Subscores.email_address`
+
     .. attribute:: ip_tenure
 
       The risk associated with the issuer ID number on the IP address. If
       present, this is a value in the range 0.01 to 99.
 
       :type: float | None
+
+      .. deprecated:: 1.8.0
+        Deprecated effective August 29, 2019. This subscore will default to 1
+        and will be removed in a future release. The IP tenure is reflected in
+        the overall risk score.
+
+      .. seealso::
+        :py:attr:`minfraud.models.Score.risk_score`
 
     .. attribute:: issuer_id_number